+++ commerce/delivery/src/main/java/ru/yandex/practicum/service/DeliveryService.java
@@ -4,6 +4,7 @@ import ru.yandex.practicum.delivery.dto.DeliveryDto;
 import ru.yandex.practicum.order.dto.OrderDto;
 
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface DeliveryService {
@@ -16,5 +17,5 @@ public interface DeliveryService {
 
     void failDelivery(UUID orderId);
 
-    Double calculateDeliveryCost(OrderDto orderDto);
+    BigDecimal calculateDeliveryCost(OrderDto orderDto);
 }
