+++ commerce/order/src/main/java/ru/yandex/practicum/service/PrepareOrderServiceImpl.java
@@ -20,6 +20,7 @@ import ru.yandex.practicum.warehouse.client.WarehouseClient;
 import ru.yandex.practicum.warehouse.dto.AddressDto;
 import ru.yandex.practicum.warehouse.dto.AssemblyRequest;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -63,12 +64,12 @@ public class PrepareOrderServiceImpl implements PrepareOrderService {
     @Override
     public OrderDto payOrder(UUID orderId) {
         Order order = getOrderById(orderId);
-        double productCost = paymentClient.getProductCost(orderMapper.map(order));
-        double deliveryCost = deliveryClient.calculateDeliveryCost(orderMapper.map(order));
+        BigDecimal productCost = paymentClient.getProductCost(orderMapper.map(order));
+        BigDecimal deliveryCost = deliveryClient.calculateDeliveryCost(orderMapper.map(order));
         order.setDeliveryPrice(deliveryCost);
         order.setProductPrice(productCost);
         log.info("order after setting productPrice: {}", order);
-        double totalCost = paymentClient.getTotalCost(orderMapper.map(order));
+        BigDecimal totalCost = paymentClient.getTotalCost(orderMapper.map(order));
         order.setTotalPrice(totalCost);
         PaymentDto paymentDto = paymentClient.createPayment(orderMapper.map(order));
         order.setPaymentId(paymentDto.getPaymentId());
@@ -106,7 +107,7 @@ public class PrepareOrderServiceImpl implements PrepareOrderService {
     @Override
     public OrderDto calculateTotalPrice(UUID orderId) {
         Order order = getOrderById(orderId);
-        double totalCost = paymentClient.getTotalCost(orderMapper.map(order));
+        BigDecimal totalCost = paymentClient.getTotalCost(orderMapper.map(order));
 
         return orderMapper.map(orderService.setTotalPrice(orderId, totalCost));
     }
@@ -114,7 +115,7 @@ public class PrepareOrderServiceImpl implements PrepareOrderService {
     @Override
     public OrderDto calculateDeliveryPrice(UUID orderId) {
         Order order = getOrderById(orderId);
-        double deliveryCost = deliveryClient.calculateDeliveryCost(orderMapper.map(order));
+        BigDecimal deliveryCost = deliveryClient.calculateDeliveryCost(orderMapper.map(order));
 
         return orderMapper.map(orderService.setDeliveryPrice(orderId, deliveryCost));
     }
