+++ commerce/payment/src/main/java/ru/yandex/practicum/controller/PaymentController.java
@@ -11,6 +11,7 @@ import ru.yandex.practicum.payment.client.PaymentClient;
 import ru.yandex.practicum.payment.dto.PaymentDto;
 import ru.yandex.practicum.service.PaymentService;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Validated
@@ -28,7 +29,7 @@ public class PaymentController implements PaymentClient {
 
     @Override
     @PostMapping("/totalCost")
-    public Double getTotalCost(@RequestBody OrderDto order) {
+    public BigDecimal getTotalCost(@RequestBody OrderDto order) {
         return paymentService.getTotalCost(order);
     }
 
@@ -40,7 +41,7 @@ public class PaymentController implements PaymentClient {
 
     @Override
     @PostMapping("/productCost")
-    public Double getProductCost(@RequestBody OrderDto order) {
+    public BigDecimal getProductCost(@RequestBody OrderDto order) {
         return paymentService.getProductCost(order);
     }
