+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -5,17 +5,18 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.order.dto.OrderDto;
 import ru.yandex.practicum.payment.dto.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
 
     PaymentDto createPayment(OrderDto order);
 
-    Double getTotalCost(OrderDto order);
+    BigDecimal getTotalCost(OrderDto order);
 
     void paymentSuccess(UUID orderId);
 
-    Double getProductCost(OrderDto order);
+    BigDecimal getProductCost(OrderDto order);
 
     void paymentFailed(UUID orderId);
 }
